Update Documentation for Active Header Highlighting
Yes, I know I call it Active Header Highlighting in some places, and Current Header Highlighting in others.

Basically the user experience I use Current, but within the code I use Active.  That's because I think "Current" is a more understandable term to end users.  But Active it better in the code, but current is used within the implementation and would be confusing.

If you are reading this and have a suggestion to improve it, I would love to hear it!